APIStage and SubDomains ENUMS moved from APIClient into EnumHelper + assignment in request object now
